refactor(x/ecocredit) migrate ecocredit utils to core
