+++ x/ecocredit/server/utils/utils.go
@@ -15,7 +15,7 @@ import (
 // GetCreditTypeFromBatchDenom extracts the classId from a batch denom string, then retrieves it from the params.
 func GetCreditTypeFromBatchDenom(ctx context.Context, store ecocreditv1.StateStore, k ecocredit.ParamKeeper, denom string) (core.CreditType, error) {
 	sdkCtx := types.UnwrapSDKContext(ctx)
-	classId := ecocredit.GetClassIdFromBatchDenom(denom)
+	classId := core.GetClassIdFromBatchDenom(denom)
 	classInfo, err := store.ClassInfoTable().GetById(ctx, classId)
 	if err != nil {
 		return core.CreditType{}, err
